sort communities
