+++ src/Pages/Communities/Community/CommunityComponents/index.ts
@@ -1,0 +1,33 @@
+// Barrel export file for all Community components
+// This simplifies imports throughout the application
+
+// Types
+export * from './types';
+
+// API Service
+export { api, ApiService } from './ApiService';
+
+// Loading Components
+export { UniversalLoader, ActionLoadingModal } from './LoadingComponents';
+
+// Modal Components
+export { ErrorAlert, ConfirmationModal } from './ModalComponents';
+
+// Filter and Avatar Components
+export { 
+  FilterDropdown, 
+  ProfileAvatar, 
+  getUserInitials 
+} from './FilterAndAvatarComponents';
+
+// Members Modal
+export { MembersModal } from './MembersModal';
+
+// Community Card
+export { CommunityCard } from './CommunityCard';
+
+// Post Components
+export { PostCard, ChatPostCreator } from './PostComponents';
+
+// Community and Post Modals
+export { CommunityModal, PostModal } from './CommunityPostModals';
